Patch by ldohrmann
[FEATURE] Added a feedback system that asks the user to participate in
an online survey. The interval in which the feedback reminder is shown
is customizable or can be disabled completely in the Preferences:
Window -> Preferences -> Saros -> Feedback

[FEATURE] A wizard to configure Saros settings is shown when the plugin
is installed or updated and some user preferences are missing.

[FEATURE] Manually connecting from the Roster View is now always
possible. If configuration details for Jabber are missing a 
wizard is shown to enter them right away.

[FEATURE] Started to developed a StatisticManager which can gather information about the session. Currently only data about the number of sessions is gathered and not transmitted at all.

[TASK] Extend the StatisticManager to gather more information and
implement a way to submit the the data to our server if the user agreed
to this.

[FIX] #2767873: Saros views not visible after plugin installation

[FEATURE] Added a check button in the invitation dialog that lets the
user specify whether he wants the dialog to close automatically if all
invitations are completed or not.
 
[FIX] #2596378: Autoclose of invitation dialog

[API] Moved access to some preferences to PreferenceUtils.

[NOP] Added Saros-Logo as an icon to our configuration dialog