[FIX/INTERNAL] Interruption stack trace when establishing SOCK5 bytestream connections
Detailed:

With SOCKS5, both peer try to establish a connection. When achieving a working SOCKS5 bytestream connection the establishing of the other one gets cancelled if still in progress. However, SMACK has no support for this interruption - the exception is caught causing a strack trace to standard output. This patch uses a separate Thread to wait until connection establishment and close it afterward.