[FEATURE/FIX] Fallback mechanism for VCS activities.
When the host has VCS support enabled, VCSActivities are sent the other participants to replicate any VCS operations like switch or update. This requires the participant to have the Team provider (Subclipse) installed. If this is not the case, trying to execute the VCSActivity resulted in an error message, and the project would get out of sync.

This patch fixes this by collecting all the other activities that are part of the VCS operation (e.g. all the file content modifications when a project is updated), and sends them as part of the VCSActivity. If the VCSActivity could not be executed because the Team provider is missing, the other activities are executed as a fallback instead.

Signed off by: Patrick Bitterling, timeout