[FEATURE] Removed "Share with VCS support..." action, added a setting "Disable VCS support" in the Prefs, default is false.
This setting can also be set in the incoming invitation dialog. If the local user has VCS support disabled, VCS operations are completely ignored by Saros (local or remote), whether she's the host or not. If e.g. just one of three participants in a session has problems with the repository, the others could still use it with no real disadvantage.

Signed-off by: patbit, donut87