[STF] define the class CommonWidgetsWrapper, so that you can access services defined on the server with e.g. alice.commonWidgets().view("Saros Buddies").bot().tree().selectTreeitem("buddies", "alice_stf@saros-con.imp.fu-berlin.de"); alice.commonWidgets().shell("Preferences").bot().tree().selectTreeitem("Saros");
Such method-chain make tests very readable and easy to write.
with view(..).bot() you can access the content of a view/shell, e.g. selectTreeItem in the view and click button "in" the view
with view(..) you can e.g. close/activate/open view and click toolbarButtons "on" the view