[STF] realize clicking sub contextMenus with e.g. alice.commonWidgets().view("Package Explorer").bot().tree().selectTreeitem("Foo_bar").contextMenu("Saros", "Share project...").click()
Such method-chain make tests very readable and easy to write.
with view(..).bot() you can access the content of a view/shell, e.g. selectTreeItem in the view and click button "in" the view
with view(..) you can e.g. close/activate/open view and click toolbarButtons "on" the view