The balloon notifications now mimics the behaviour of windows balloons more closely. I.e. the balloons are closable by clicking anywhere on them (i also corrected the formatting, the nasty gray background is gone).
Unfortunately there seems to be a bug in the BalloonNotification implementation (External Library) that prevents them from being closable by clicking anywhere, when the closing X is shown. So right now there is no closing "X".
If a user is annoyed by the balloon he/she will click on it to try to get rid of it by left or right click, successfully closing it, and won't be bothered by not having a sophisticated close button.