Refactored the BaseResourceSelectionComposite to be more efficient. Removed unused methods, removed redundant resource wasting (redundant method calls, inefficient methods). Refactored the invitation wizard so it does not take ages (without feedback) before opening. The user now gets to see the wizard instantly, but blocked until the initial selection which is taken from the current perspective is fully applied.
The biggest speed improvement came by removing the redundant collecting of all selected elements just to see if there are 0 or at least one element selected (sic!). This was taking about 80% of the total time for work done in the wizard.
The removede "SetChildrenUngrayed" is done by the checkboxTreeViewer itself when setting checkstates of the parent objects - it turned out it's not necessary to do that manually. Sorry i didn't mention that.

new method "hasSelectedReresources" should be used in cases one just needs to know if there is at least ONE selection.
If no object is selected, there are also no grayed elements. If one object is selected, it does not matter if anything is grayed or not, there is at least ONE thing selected. So no special filtering is required.