[NOP] Use American English in comments and code
The Guideline requires to use 'English' but does not specify the
flavor that should be used. This leads to a mix of American and
British English. The programming language used and the IDE targeted
default to American English, do so as well.

colour => color
behaviour => behavior (a package needs to be renamed too)
synchronise => synchronize

Change-Id: Ia2ba26d0e41fc9f87670a9d27c1d6b948bafaa19