[CHANGE] Remove broken and unused revert handling code
The revert handling code got disabled in SVN revision r2969 when adding
the possibility to add projects to an existing session. The attempt to
re-enable this change leads to modification circles. In case Bob is
following Alice and Alice is closing an editor, the code will trigger
on both Alice and Bob, Alice and Bob will now respond to the
modification each other made and will continue to do so.

The right behavior for following should be discussed because right now
both Alice and Bob will be asked if they want to save the modification.

Change-Id: If763ffe419838bc0ff5574114bfcc642f52b7bc4