[INTERNAL] The StopManager should exist within a SarosSession
The code currently installs a listener on the SarosSessionManager to
find the start and end of a SarosSession and calls reset at the end
of each SarosSession. This is a clear indication that the StopManager
is not global. It should at least be scoped by SarosSession maybe even
smaller. This change removes the IllegalStateException during the
InviteAndLeaveStressTest due working on different StartHandles.

Change-Id: I91c95dfcbc32abdda1cee3eb9985240906722809