[INTERNAL] Remove the unused streaming archive feature
It appears that this feature was never finished and features
versious design and implementation issues. E.g. the usage of
the IProgressMonitor and cancellation is broken. From a design
point of view one needs to wonder why this is not done by
subclassing the negotiation class.
This commit changes the marshalling of the ProjectsAddedActivity,
I am not sure which version number to increase now.

Change-Id: If43375a0efb217ba7696f3c2c1db6cf082450e56