[INTERNAL] Address ConcurrentModificationException in StopManager.stop
The code currently tries to exit as soon as the doneSignal has
timedout and monitor.setCanceled has been called once. The code
will then restart already stopped users. What happened is that
during this loop another StartHandle is added to the list causing
a ConcurrentModificationException. This can happen during
the invitation stress test. The race condition comment should be
addressed now. The cancellation exception is either thrown inside
the stop method or not thrown at all. Document the usage of the
wait condition and that multiple threads might wait for the signal.

Change-Id: Ic5c757b10c61456ce679aa30a5fd17a904f13239