[INTERNAL] Fix regression of StoppedAction being enable dby default
This was broken in r3812 and spotted by Stefan Rossbach. The button
is enabled by default, before the above change registering with the
StopManager.getBlockedObservable() lead to the setUnblockEnabled being
called. Address the issue by disabling the button by default. I decided
to not call setUnblockEnabled as it runs SWT async.

Change-Id: If25cb2d2e8e0b077585fa8a2b7dbe46712333cbe