[FIX] Remove "Stop running process" icon - ID: 3500370
Disabled the visibility of this icon when not in debug mode. Currently it only shows that
the user is blocked. If you press on this icon, nothing happened which is very confusing to
the user. The icon itself also does not represent any visual state that indicated that the 
user cannot type any data into the opened editors.