The remoteProgressMonitor now enables the host to see the progress of an SVN checkout of an invited user. The host can also cancel the svn checkout, and so remains in control of the session. This increases progress reporting drastically (before: no reporting at all, progress stops in "sending file list" for about 15 minutes (svn checkout of saros)), now: cancelable progress monitor appears.
I cannot display an estimation on how long this will take, but the users usually know that when the svn checkout reaches file xy (e.g. the libs on saros), it will take just a few minutes until it is finished, so this helps the user determine if he needs yet another cup of coffee.