The buddySessionDisplayComposite now display who is following who as a child tree element of the user element. Also shows the currently open file (if any) as another tree element for each user.
Doubleclick on the filename "jumps to position".
Double click on the user toggles follow mode.

Also indicates which user is the HOST, because users could not remember which user was the host, to change session rights (grant write access) > Heuristic: Recognition rather than recall

We want to get rid of the host role, but this won't happen any time soon. If we introduce some sort of host role transferring, we especially need to know who is the host.

Also sorts the userlist so "YOU" is always at the bottom (the item with the least user interaction should not be somewhere in the middle, randomized)...

Also has some nicer formatting for the colors of the users.