I removed the getting started wizard, which is violating lots of guidelines and added  an internal browser window (as editor part) with the page (which i updated) http://www.saros-project.org/GettingStarted?template=viewprint that contains the tutorial on how to use saros.
This way the tutorial is unobstrusive, can directly provide more information, can contain links to features so users get to know saros better, etc. We can even show a video introduction later.

The saros tutorial will open as long as saros is not configured properly (no xmpp account). The user can open the tutorial at anytime by using the "Getting started" menu item.

The configuration wizad is now MODELESS, so the eclipse window and the tutorial can be accessed even when it automatically opens.
This is a temporary fix, we need to change the saros view to prominently show a button to open the config wizard, and hide everything else in the saros view, until the user has set an xmpp account.