The ResourceSelectionComposite gets an Undo/Redo functionality to avoid frustration when accidentally clicking on a project checkbox after setting a complicated partial sharing selection (thus losing it).
The composite also gets a new feature to remember selections with names and restore those (and edit/delete them too).
When finishing the invitation wizards, a default stored selection is updated "-Last selection-" which can be restored with one click in the wizard, enabling faster "resuming" of the previous session.

Also includes a Bugfix for the problem that the wizard did not automatically select a newly created project (from within the wizard). The NewProjectListener does not work properly, so i implemented it in another way (probably more efficient also).