[INTERNAL][nwarnatsch] fix the cancellation during an incoming project negotiation
In the IncomingProjectNegotiation, the projects are added to the ResourceMapping before the archive is received. 

The user cancels the incoming-process. Now, the method executeCancellation closed the session just if the ResourceMapping is empty. That couldn't be, because it was filled at the start of the process.

This fix removed the projects, which arrived during the IncomingProjectNegotiation, from the ResourceMapping if the cancellation is executed

[reviewed by: Holger Freyther and Stefan Rossbach]