[INTERNAL] ISarosSession is no IActivityListener anymore.
The IActivityListener can be installed on the IActivityProvider
and will be notified about the creation of activities. This design
was broken early on and lead to code that directly calls the
method on the ISarosSession. Fix the design by making the SarosSession
own an IActivityListener.

Restoring the initial design is helpful because we have a few
real listeners around the EditorManager and it is easier to restore
the design than to introduce coupling.

The next step is to remove the direct sending of activities.

Change-Id: Ie05ed9599e5eebcad57d1fdd75879475fc9ff638