[INTERNAL] Fix NPE in the StopManager code path
The StopManager was created before SarosSession's internal activity
listener was created. The StopManager constructor will register with
the SarosSession and an empty ActivityListener will be installed that
will lead to the NPE. Allocate the StopManager _after_ the activityListener
has been created. A picocontainer for the SarosSession might solve these
dependency issues in a more generic way.

Change-Id: I2a30b7f290e992c8d9ecc1536ffc9c5d0dba9583