[INTERNAL] removed broken session resuming
The feature was never fully implemented.
In it is current state it started to queue messages
that could not been send. This queue was never read
again. This forces a user to stay forever in the session
he is no longer part of. The user is now removed from
the session if sending data to him fails.