[INTERNAL] Remove the IProjectActivityDataObject interface.
The IActivityDataObject is an intermediate interface to ease with
serialisation. It should only be used right before we send the message
or right after we have received it. Remove additional interfaces
as they should not be used/called.

Move the getProjectID into the EditorActivity as it is only used
by the SarosSession on the EditorActivity.

Change-Id: I3914cf41d5cb4b49e18cb9b6f4603d3d8319a439