[INTERNAL] Introduce a JUnit test for the StopManager
Create a JUnit test for the StopManager/StartHandle that goes
through some easy cancelation scenarioas, verifies the messages
sent for lock/unlock (but not their acks) and makes two
StopManagers communicate with each other. The multi user stop
can not be tested due trying to run code on the SWT Thread.

I didn't check how much the code coverage of this class increases.

Change-Id: I236c245f80ff82acd9dd39bef22d03a90f0899d1
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/24
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>