[INTERNAL] The JUnit tests run without GUI, do not synchronize on the SWT Thread
Our test infrastructure does not initialize SWT when running the
JUnit tests, internal parts of Saros should not know about SWT
either. Remove the need to call IActivityListener.activityCreated
from the SWT thread and move the synchronisation to the listener,
update the users of the Blockable listener to synchronize in the
listener as well.

Remove the Utils.runSafeSWT* calls and re-introduce them in the
SarosSession and the UndoManager ActivityListeners.

Change-Id: I537a51bc19f0c5bf7b76365a2e5afb0e7fea8380
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/25
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>