[INTERNAL] Create a child container inside the Session and test
A lot of objects have a lifetime that is limited by the session
but use the ISarosSessionListener to start/stop and reset the
state. Create a PicoContainer per SarosSession as a child container
of the SarosContext and move the StopManager into this container.

Add a JUnit testcase to test the creation, starting, stopping and
disposing of the session. This is done by creating the ISarosContext
interface, adding the methods used by SarosSession to that interface
and by making the ActivitySequencer use the ISarosContext of the
session.

Most code like this can be migrated to use the Startable interface
of the picocontainer to implement the lifecycle:

    protected ISarosSessionListener sessionListener = new AbstractSarosSessionListener() {

        @Override
        public void sessionStarted(ISarosSession newSarosSession) {
            // save the id of the just started session
            currentSessionID = sessionID.getValue();
        }

        @Override
        public void sessionEnded(ISarosSession oldSarosSession) {
            if (isErrorLogSubmissionAllowed())
                submitErrorLog();
        }

    };

migrate to:

Class implements Startable {
    void start() {
        // action on session start
    }

    void stop() {
        // action on session stop
    }
}

Change-Id: I79bcf2a502efa49db26501bcd3ec6772836d7a22
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/26
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Holger Freyther <holger@freyther.de>