[INTERNAL] Kill the flushTask using a poison pill
The SarosSession create/start/stop/dispose infinite test fails
with an OutOfMemory Exception because the ActivitySequencer timers
continue to run after the end of a session. The current comment shows
that this problem is known and added a 30 second timeout. This commit
uses the poison-pill approach to inject the shutdown into the queue.

Change-Id: Ia80b584c5e9f14e20acf49fb011f31aac68ce2db
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/30
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>