[INTERNAL] Move the feedback managers and collectors into session
The collectors are clearly scoped by the session, this can be seen
by the abstract "clearPreviousData" method. Make all collectors use
the ISarosSession instead of the ISarosSessionManager, make sure to
remove installed listeners at the end of the session. Remove the usage
of the StatisticManager from outside the SarosSession and replace
it with static methods (this code can be move to a new class now).

Change-Id: I69b86582b0a61803eb5ebbcb46f0ff39f7be4b81
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/36
Tested-by: Jenkins CI
Reviewed-by: Holger Freyther <holger@freyther.de>