[INTERNAL] Add unittest for the Collectors creation/destruction
Add a test that fills a picocontainer with the collectors, then starts
it, verifies that all available collectors became active, verifies that
at the end of the session all of them are inactive again and that all
installed listeners have been removed.
The mocking code could move to a different helper/support package as
more tests could make use of them.

Change-Id: I08867628950c01de366f68620f47b2d631541e3c
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/41
Tested-by: Jenkins CI
Reviewed-by: Holger Freyther <holger@freyther.de>