[INTERNAL] Fix NPE and frequency of scheduling the flushTask timer
The BlockingQueue.poll/Queue.poll is not a blocking operation and
will return null that will cause a NPE later in this code. Use the
blocking .poll() with timeout again, add a null check to see if the
polling has timed out. This time the InterruptedException is handled
in the caller. This change also changes the frequency this timer task
is scheduled.
The NPEs were found when browsing alice.log of the Saros STF Job on
our Jenkins installation, they have not caused any failure.

Change-Id: I8f766a6eee7790b6056ad1b1a0764b2a80970669
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/65
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>