[INTERNAL] Address NPE in the HostLeftAloneInSessionHandler
Stefan has already deprecated this method and this exception is one of
the reasons that this was a good idea. Add some more documentation about
what to do with this method and how to avoid it in new code.

The HostLeftAloneInSessionHandler  could be fixed in a various ways,
e.g. one could use the SarsoSession inside the user that has left as
this may not be null. I decided against doing that and added a null
check as I was afraid that the code could attempt to leave an already
dead session.

java.lang.NullPointerException
	at de.fu_berlin.inf.dpp.ui.eventhandler.HostLeftAloneInSessionHandler$1.userLeft(HostLeftAloneInSessionHandler.java:45)
	at de.fu_berlin.inf.dpp.project.internal.SharedProjectListenerDispatch.userLeft(SharedProjectListenerDispatch.java:32)
	at de.fu_berlin.inf.dpp.project.internal.SarosSession.removeUser(SarosSession.java:587)
	at de.fu_berlin.inf.dpp.net.business.LeaveHandler$Handler$1.run(LeaveHandler.java:86)
	at de.fu_berlin.inf.dpp.util.Utils$6.run(Utils.java:338)

Change-Id: I3872814e1f5deef952b89843acf0b59d3b775aff
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/69
Tested-by: Jenkins CI
Reviewed-by: Norman Warnatsch <norman@warnatsch.name>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>