[FIX] Fix StopManager.stop exception occuring during STF tests
It is possible that users disappear from the session during the
stop handling. It occured during the STF testing and this adds
a testcase and a fix for that behavior. This only tests that a
user is leaving after the stop has been initiated, it is likely
that there might be issues if a user manages to leave more earily
during the project negotiation.

java.lang.IllegalArgumentException: StopActivity contains recipient which already left: StopActivity (id: 0614076147051098924886315126, type: UNLOCKREQUEST, state: INITIATED, initiator: jenkins_alice_stf, affected user: jenkins_carl_stf, src: jenkins_alice_stf)
	at de.fu_berlin.inf.dpp.synchronize.StopManager.fireActivity(StopManager.java:439)
	at de.fu_berlin.inf.dpp.synchronize.StopManager.initiateUnlock(StopManager.java:424)
	at de.fu_berlin.inf.dpp.synchronize.StopManager.resumeStartHandle(StopManager.java:515)
	at de.fu_berlin.inf.dpp.synchronize.StartHandle.start(StartHandle.java:66)
	at de.fu_berlin.inf.dpp.synchronize.StopManager.stop(StopManager.java:263)
	at de.fu_berlin.inf.dpp.invitation.OutgoingProjectNegotiation.createProjectArchives(OutgoingProjectNegotiation.java:504)
	at de.fu_berlin.inf.dpp.invitation.OutgoingProjectNegotiation.start(OutgoingProjectNegotiation.java:124)
	at de.fu_berlin.inf.dpp.project.SarosSessionManager$OutgoingProjectJob.run(SarosSessionManager.java:653)
	at org.eclipse.core.internal.jobs.Worker.run(Worker.java:54)

Change-Id: I11619163ffec97b358aec03719255a400ba114a7
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/74
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>