[INTERNAL] Verify the removal of all ActivityProviders
Verify that before a session starts no activity providers are
registered with the session, that after start there are activity
providers and that after the session has been stopped one is back
to zero. I picked the idea from saros-light, the change is a bit
different though and the test is only present here.

Change-Id: I8aa20fc68910af682384e319a3a4d7674e98615c
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/76
Tested-by: Jenkins CI
Reviewed-by: Holger Freyther <holger@freyther.de>