[INTERNAL] It is odd to use someone elses logger, stop that
The ProjectDeltaVisitor was added in 3af292ec90acb95e75e6f7
(svn revision=2344) and consisted out of code being moved from the
SharedResourcesManager into the ProjectDeltaVisitor. The
ResourceActivityFilter was created in ed44f2c2038610e81c1f1
(svn revision=2784) and involved copying code from the
ProjectDeltaVisitor. I assume the access to the
SharedResourcesManager.log was a left over and this commit is
cleaning it up.

Change-Id: I464a6d0a7d86f441f0a258707b0eed78bea4852e
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/79
Tested-by: Jenkins CI
Reviewed-by: Holger Freyther <holger@freyther.de>