[FIX] Fix STF failures after the FeedbackManager/ErrorLogManager move
The ErrorLogManager/FeedbackManager/StatisticManager code has been
used outside the session for configuration but once for uploading
data. Make the configuration code static (like it was done in the
previous commit) and remove the upload error log outside of a session
feature. For the uploading code we will need to remember the last
session id to continue to provide this feature.

For both the ErrorLogManager and the FeedbackManager we will need
to extract the static configuration into a *Configuration class.

Change-Id: Ie5d275f44d9e461a077f3e69f97da80ebd741ac6
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/88
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>