[INTERNAL] Remove the IProjectActivityDataObject interface.
The IActivityDataObject is an intermediate interface to ease with
serialisation. It should only be used right before we send the message
or right after we have received it. Remove additional interfaces
as they should not be used/called.

Move the getProjectID into the EditorActivity as it is only used
by the SarosSession on the EditorActivity.

Change-Id: I1c99075d1b9fd105f1c72b76327b7a04bc7c6af7
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/91
Tested-by: Jenkins CI
Reviewed-by: Norman Warnatsch <norman@warnatsch.name>