[INTERNAL] Remove user in session check from fire activity
The user can leave the session at any time. This can happen between
the check and the fireActivity method but also after the check in
fireActivity and super.fireActivity. Accept that this test can never
function properly and remove it.
It is possible that the StopManager sent a request and will wait for
the ACK forever. It needs to be specified how this can be properly
handled, e.g. should this be siltenly ignore, should it throw an
exception?

Change-Id: Ia50103f45dc0636ee2ec0c929d57565e2936047e
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/98
Tested-by: Jenkins CI
Reviewed-by: Holger Freyther <holger@freyther.de>