[PERFORMANCE] reduced size of marshalled objects
This optimization was lost when restoring the old network
layer. XStream by default apply pretty printing when
marshalling objects to XML. This is not necessary
because the data is not presented to the user and
is only processed by the application. This change
reduced the size of marshalled objects by
10% on average.


Change-Id: Icd3357cdbe3b47a22ce1f1def8aa61067596bd05
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/138
Tested-by: Jenkins CI
Reviewed-by: Holger Freyther <holger@freyther.de>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>