[INTERNAL] Remove sending 'small' messages via the chat.
The DataTransferManager has an IBB and Socks5 transport. The
XMPPTransmitter uses the DataTransferManager to send the data.
In case the transmit failed it will be sent by chat or retried.
If retrying has failed too many times the DataTransferManager
will be requested to use IBB for further transfers.

The secret of the XMPPTransmitter and DataTransferManager are
mixed. This attempts to simplify the code a bit by removing the
fallback to send the message using chat. In case IBB|Socks5 does
not work reliable it is better to have the communication fail
early and fast.

Change-Id: Iad33ffe0dc109803234910f36ffd1a3f188e47fe
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/148
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>