[INTERNAL] Use the ITransmitter interface and remove unused API
Make IncomingSessionNegotiation, OutgoingSessionNegotiation and
InvitationHandler use the ITransmitter interface instead of the
XMPPTransmitter implementation. Add required methods to the
ITransmitter interface. Remove unused functions from the interface
and make implementations private/protected.

The extensions are a secret of using XMPP/Smatch for the transport,
do not leak this knowledge into other packages. Move the usage of
extension providers into the XMPPTransmitter class. Instead of
creating the providers on the fly cache them in the SarosContext
like it is done for the other extension providers.

The whiteboard requires the XMPPTransmitter to keep three methods
of the API. installReceiver, receive and sendToProjectUser.

Change-Id: Ic55aab0b45736a67cdd9377e726407bb5a0682b1
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/151
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Tested-by: Jenkins CI