[FIX] Use asynchronous processing for incoming activities
Performance improvement by asynchronous processing of incoming
activities. The activityDispatcher was deleted, because it worked under
unix-systems too slow. Now, the incoming activities will be passed to a
method, that starts asynchronous a runnable in the swt-thread. That
patch is a performance patch for UNIX-systems. On windows, it won't
influence the performance.

Change-Id: I768a63e3bae5dd540ac7a5e52373812c81be88e4
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/158
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>
Reviewed-by: Sebastian Starroske <sebastian.starroske@googlemail.com>
Reviewed-by: Norman Warnatsch <norman@warnatsch.name>