[STF API] extended API to handle Idaede3daa2c12d2502b0d2a6309f826efd861a16 commit
public void setMessage(String newMessage, int newType)

Sets the message for this dialog with an indication of what type
of message it is.

    The valid message types are one of NONE, INFORMATION,WARNING, or
ERROR.

    Note that for backward compatibility, a message of type ERROR is
different than an error message (set using setErrorMessage). An error
message overrides the current message until the error message is
cleared. This method replaces the current message and does not affect
the error message. 

Change-Id: I81f6eabd0b119f2694113bc9f6ce8d96a9c61a0c
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/214
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>