[STF] fixed test case InviteAndLeaveStressTest
Commit 834005831eb5b0d6ecaf0b45dddcec8d142ffff3
introduced a bug fix for the "Add Projects to Session"
wizard which will now override the destination project
name.

The test case relied on the fact, that it could
add as many projects as it wants, because new auto
generated names were used.

This fix generate the projects first, and then shares
them will fully qualified names.


Change-Id: Idc8fae4ecd00303480264cc90e0b59f37c6ec52e
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/223
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>