[UI] Default color attribution for Saros annotations - part 1
The user color is now the default color for contributions.
The selection annotation uses a slightly lighter color to
distinguish between the two. This lighter color gets represented in the
chat as well.
What remains the same is that contributions will be displayed in a
darker color than the selections.

(Before: User color was the default for selections. Contribution
color was slightly darker.)

This UI change addresses the following problems:
1) usability problem "low contrast between the selection and
contribution color"
Users had a hard time to see the cursor or highlighted text (selection)
within the last editor changes (contribution).
2) correct calculation of lighter/ darker colors
The lightness of colors will now be calculated correctly by using the
HSL color model. For this purpose the already existing method
"ColorUtils.addLightness" is used (see de.fu_berlin.inf.nebula).
3) aesthetics
From an aesthetic point of view Julia, Franz and me decided to
brighten up the default colors - defined in the plugin.xml.


Change-Id: If549225763668e99c054c565c41b333ae575190d
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/250
Tested-by: Jenkins CI
Reviewed-by: Norman Warnatsch <norman@warnatsch.name>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>