[FIX] #3514675 Useless notifications for 3rd person joining session
"Buddy joined the session" notifications are only displayed for users
which have joined the session after the local user. Before this fix,
the local user was notified that a buddy had joined the session, even if
this buddy was already in the session. For example, Alice and Bob are in
a session and Carl is invited, Carl is notified that Bob has joined the
session.

Copied from SWTP-2012 repository.
 - Commit 1f883535e6a8928c8edcc77e10be6857e47734f0
 - Reviewed on http://saros-build.imp.fu-berlin.de/gerrit/251

Change-Id: I87d5df55957305feb478cb939efb906bac0a8e46
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/264
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>