[PRODUCTION] reduced memory size for Findbugs task
Added a hardcoded limit for Findbugs on how much memory
it should use. The file is also present in the local
production environment but the Saros job is the only
job that execute Ant on the build file from the SCM.

All other job uses the local version. This should
be changed too !

Change-Id: I8d175f0dbac2e4a255de8ad0777e0f617b36c4d8
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/294
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Tested-by: Jenkins CI