[INTERNAL] FileZipper is not dependent on a monitor anymore
The methods zipFiles(..) and createProjectZipArchive(..) of the util
class FileZipper were using monitors which were passed as arguments.
Now those methods expect a ZipListener instead of a monitor. This
ZipListener is notified when a new zip entry is created and after a
chunk of data has been deflated.

This helps separating the zipping (which is the main intention of those
functions) from the updating / handling of the monitors. Besides it
allows other implementations of the ZipListener.

Change-Id: I80ca86c1bf467961e8f343183e83e8b60069151b
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/277
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>