[UI] Disable left Viewport Annotation by default
The left viewport annotation is not displayed by default anymore,
because it is confusing for users - and us, too.

Interviews showed that more users take notice of the right viewport
annotation and are more likely to use this one. They do recognize the
left viewport annotation, too but don't use it. Furthermore it is still
possible to enable it via the Eclipse preferences "General > Editors >
Text Editors > Annotations".
Other software like SubEthaEdit show their viewport annotations on
the right handsight too.

So even though the left viewport annotation contains some really useful
information, users are not able grasp it. Since there is no real
solution for this dilemma it is better to go for an "simplicity"
instead of a "quantity of functionality" approach.

Change-Id: I84b2d87859981a8a153ae711cde5fbdfd555ce27
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/290
Tested-by: Jenkins CI
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>