[FEATURE] Improve title of single and multi user chats
If there is a user name available for the contact of a single user chat,
that is used as title. Otherwise the bare JID is used.
For multi user chats (session chat), the usual 'Session Chat' title is
returned for now, as it is not possible to create a multi user chat
despite the session chat.

Copied from SWTP-2012 repository.
 - Commit a13c17b0e5aaa562658943411d2485bce502f7cd
 - Reviewed on: http://saros-build.imp.fu-berlin.de/gerrit/285
 - Reviewed by: Nils Bussas <nils.bussas@gmail.com>
 - Reviewed by: Arsenij Solovjev <xeper000@gmail.com>
 - Reviewed by: Michael Scheppat <mschepp@zedat.fu-berlin.de>
 - Reviewed by: Vera Schlemm <vera.schlemm@fu-berlin.de>

Change-Id: I2b01561ebb5a568469fdbba2bc53752d3aee3acc
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/352
Reviewed-by: Franz Zieris <zieris@inf.fu-berlin.de>
Tested-by: Jenkins CI