[FIX] #3458952 File changes during Invitation corrupts session
While inviting Carl, Alice got the need based dialog for every key
inserted by Bob.

To fix this, the feature "Pre-Sharing" was removed. Pre-Sharing means
that during the invitation process the files currently opened/edited by
the host are sent directly to the invitee. This enables the invitee to
start working before the complete archive is received.

This feature was removed, because it makes the invitation process
unstable.

Change-Id: I72188db29f69adeebef4e31fac53f0379b95da14
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/328
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>
Tested-by: Jenkins CI