[FIX] #3512780 Getting Started not noticeable enough
The wizard is no longer automatically opened. Along
with commit 97ba6a5c002dc2f47e7eb5e3c367b0217e799591 the
user will now be able to easily access the configuration
wizard so there is no need to open it automatically which
may also be annoying to the user.

Change-Id: I9345c41245a9c1f85dfedbcfd5c1f54024f9b802
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/390
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>