[STF] fixed pressEnterKey method in the SarosSWTBotChatInput class
As the KeyAdapter is not noticed about special key events
like carriage return that are generated by the mock keyboard
this method simulated the pressing of carriage return by calling
methods from the logic instead.

The method that was responsible for sending messages was moved.
This patch will call the correct method again.

Change-Id: I915fb1be6f4b48bbc3a89e05f4e989a667cea1e8
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/405
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>