[FIX] #3589495 Chat display does not properly update its contents
The ChatControl class used a STATIC hash map to cache colors.
This was causing malfunction if another chat tab got refreshed
because it disposed all colors, so when you switched back to
another chat tab a repaint of the widgets would fail as they
were now using a disposed color.

The fix makes the cache local, caching only colors for a single
chat tab and not for all tabs to prevent the problem described above.

Change-Id: I6cd2941cfe0a6d4ba6743d93d7d6d262cc3273d8
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/423
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>