[FIX] do not allow a resource qualified JID to be added as a roster entry
It is currently possible to add a resource qualified JID as contact
to the roster. Saros is currently unable to delete such entries
afterwards. This is caused by the SelectionRetriever which tries
to adapt a RosterEntryElement to a RosterEntry. The adapter
will return null, as the Smack JavaDoc for roster.getEntry is
ambiguous and not very well documented. The method may not return
what you would expect when reading its documentation.


Change-Id: Iaf34b2d151e1ccf664c65723ee987d2d223bdb03
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/414
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>