[JUNIT] removed all test cases that depend on hardware
The test cases needs a full working XMPP server.
The server might not be always up.

Furthermore if two or more developers are running the
tests at the same time it is likely that failures will
occur (CI is also affected).

Regression time went down from 1 minute to approx. 6 seconds.

Change-Id: I51a38853924e2f9f707b85f2b39b12e0d7b70b1d
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/439
Tested-by: Jenkins CI
Reviewed-by: Stefan Rossbach <srossbach@arcor.de>